requirements.txt
about-time==4.2.1
absl-py==1.4.0
alive-progress==3.1.4
astor==0.8.1
attrs==23.1.0
autograd==1.6
autolab-core==1.1.1
autolab-perception==1.0.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
cma==3.2.2
colorlog==6.7.0
cycler==0.11.0
deap==1.3.3
Deprecated==1.2.14
dill==0.3.6
ffmpeg-python==0.2.0
fonttools==4.38.0
freetype-py==2.4.0
future==0.18.3
gast==0.2.2
google-pasta==0.2.0
GPUtil==1.4.0
gqcnn==1.3.0
grapheme==0.6.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.2.0
jsonschema==4.17.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lxml==4.9.2
mapbox-earcut==1.0.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
mpmath==1.3.0
multiprocess==0.70.14
networkx==2.6.3
numpy==1.21.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
psutil==5.9.5
pycollada==0.7.2
pyglet==2.0.dev22
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
PyWavelets==1.3.0
requests==2.31.0
Rtree==1.0.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
svg.path==6.2
sympy==1.10.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.3.0
tflearn==0.5.0
threadpoolctl==3.1.0
tifffile==2021.11.2
trimesh==3.21.7
typing_extensions==4.6.3
urllib3==2.0.3
visualization==1.0.0
Werkzeug==2.2.3
wrapt==1.15.0
xxhash==3.2.0
zipp==3.15.0
